refactor: entity 패키지 추가에 따른 import 구문 수정
